Reset contact_surface_layer to 0.0 m
This is what was used before; if I read the documentation correctly,
this only affects objects that are penetrating at the start of the
simulation:

https://classic.gazebosim.org/tutorials?tut=physics_params&cat=physics

At least in my tests, the wheels of the MiR penetrate the ground plane
by 0.01 m (because the MiR URDF sets minDepth to 0.01), as they should.
This is not affected by contact_surface_layer. Since there is no reason
to change it, we shouldn't.