configure physics parameters for a faster simulation by default (#3)
* configure physics parameters for a faster simulation by default

* Reset iters and sor values to previous ones for slow physics

These were the values used before for the "slow" simulation. They are
tested and working, no reason to change them; also, "slow" is just for
backup, "fast" is the new default.

* Reset contact_surface_layer to 0.0 m

This is what was used before; if I read the documentation correctly,
this only affects objects that are penetrating at the start of the
simulation:

https://classic.gazebosim.org/tutorials?tut=physics_params&cat=physics

At least in my tests, the wheels of the MiR penetrate the ground plane
by 0.01 m (because the MiR URDF sets minDepth to 0.01), as they should.
This is not affected by contact_surface_layer. Since there is no reason
to change it, we shouldn't.

* Remove real_time_factor

The real_time_factor (i.e., how much faster the simulation runs compared
to realtime) is computed as:

real_time_factor = max_step_size x real_time_update_rate

...except when real_time_update_rate is 0, then real_time_factor is
infinity (i.e., as fast as possible).

See: https://classic.gazebosim.org/tutorials?tut=physics_params&cat=physics

Setting it explicitly makes no difference and is actually confusing, so
let's remove it.

* Add some more comments

---------

Co-authored-by: Martin Günther <martin.guenther@dfki.de>